# docs/openapi.contracts.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/stalls": {
      "get": {
        "operationId": "StableDetailsController_stalls",
        "parameters": [],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List club stalls",
        "tags": [
          "stable"
        ],
        "x-implementation-status": "contract-only"
      },
      "post": {
        "operationId": "StableDetailsController_createStall",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStallDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create stall",
        "tags": [
          "stable"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/stalls/{id}": {
      "get": {
        "operationId": "StableDetailsController_stall",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get stall",
        "tags": [
          "stable"
        ],
        "x-implementation-status": "contract-only"
      },
      "patch": {
        "operationId": "StableDetailsController_updateStall",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStallDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update stall",
        "tags": [
          "stable"
        ],
        "x-implementation-status": "contract-only"
      },
      "delete": {
        "operationId": "StableDetailsController_deleteStall",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft-delete stall",
        "tags": [
          "stable"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/stalls/{id}/assignments": {
      "get": {
        "operationId": "StableDetailsController_assignments",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List stall assignment history",
        "tags": [
          "stable"
        ],
        "x-implementation-status": "contract-only"
      },
      "post": {
        "operationId": "StableDetailsController_assign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStableAssignmentDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign horse and groom to stall",
        "tags": [
          "stable"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/stable-assignments/{id}/end": {
      "post": {
        "operationId": "StableDetailsController_endAssignment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "End stable assignment",
        "tags": [
          "stable"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/horses/{horseId}/feeding-plans": {
      "get": {
        "operationId": "StableDetailsController_feedingPlans",
        "parameters": [
          {
            "name": "horseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List horse feeding plans",
        "tags": [
          "stable"
        ],
        "x-implementation-status": "contract-only"
      },
      "post": {
        "operationId": "StableDetailsController_createFeedingPlan",
        "parameters": [
          {
            "name": "horseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeedingPlanDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create feeding plan for approval",
        "tags": [
          "stable"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/feeding-plans/{id}/approve": {
      "post": {
        "operationId": "StableDetailsController_approveFeedingPlan",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve feeding plan as Trainer or Vet",
        "tags": [
          "stable"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/horses/{horseId}/checklists": {
      "get": {
        "operationId": "StableDetailsController_checklists",
        "parameters": [
          {
            "name": "horseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List horse daily checklists",
        "tags": [
          "stable"
        ],
        "x-implementation-status": "contract-only"
      },
      "post": {
        "operationId": "StableDetailsController_createChecklist",
        "parameters": [
          {
            "name": "horseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChecklistDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create assigned daily checklist",
        "tags": [
          "stable"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/incidents": {
      "get": {
        "operationId": "StableDetailsController_incidents",
        "parameters": [],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List stable incidents",
        "tags": [
          "stable"
        ],
        "x-implementation-status": "contract-only"
      },
      "post": {
        "operationId": "StableController_report",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportIncidentDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Report stable incident",
        "tags": [
          "stable"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/incidents/{id}": {
      "get": {
        "operationId": "StableDetailsController_incident",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get stable incident",
        "tags": [
          "stable"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/incidents/{id}/status": {
      "patch": {
        "operationId": "StableDetailsController_incidentStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIncidentStatusDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update incident resolution status",
        "tags": [
          "stable"
        ],
        "x-implementation-status": "contract-only"
      }
    },
...
  },
  "info": {
    "title": "Racehorse API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateStallDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          }
        },
        "required": [
          "code"
        ]
      },
      "CreateStableAssignmentDto": {
        "type": "object",
        "properties": {
          "horseId": {
            "type": "string",
            "format": "uuid"
          },
          "groomId": {
            "type": "string",
            "format": "uuid"
          },
          "startAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "horseId",
          "groomId",
          "startAt"
        ]
      },
      "CreateFeedingPlanDto": {
        "type": "object",
        "properties": {
          "effectiveFrom": {
            "type": "string",
            "format": "date"
          },
          "effectiveTo": {
            "type": "string",
            "format": "date"
          },
          "ration": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "effectiveFrom",
          "ration"
        ]
      },
      "CreateChecklistDto": {
        "type": "object",
        "properties": {
          "groomId": {
            "type": "string",
            "format": "uuid"
          },
          "checklistDate": {
            "type": "string",
            "format": "date"
          },
          "items": {
            "type": "object",
            "additionalProperties": {
              "type": "boolean"
            }
          }
        },
        "required": [
          "groomId",
          "checklistDate",
          "items"
        ]
      },
      "UpdateIncidentStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "IN_REVIEW",
              "RESOLVED"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "ReportIncidentDto": {
        "type": "object",
        "properties": {
          "horseId": {
            "type": "string",
            "format": "uuid"
          },
          "description": {
            "type": "string"
          },
          "urgent": {
            "type": "boolean",
            "default": false
          },
          "mediaAssetId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "horseId",
          "description"
        ]
      },
...
    }
  }
}